# chore: enable sqlite wal journaling (RC)