--- a/persistence/src/androidMain/kotlin/com/wire/kalium/persistence/db/GlobalDatabaseProvider.kt
+++ b/persistence/src/androidMain/kotlin/com/wire/kalium/persistence/db/GlobalDatabaseProvider.kt
@@ -15,7 +15,7 @@ import com.wire.kalium.persistence.daokaliumdb.ServerConfigurationDAO
 import com.wire.kalium.persistence.daokaliumdb.ServerConfigurationDAOImpl
 import com.wire.kalium.persistence.util.FileNameUtil
 import com.wire.kalium.util.KaliumDispatcherImpl
-import net.sqlcipher.database.SupportFactory
+import net.zetetic.database.sqlcipher.SupportOpenHelperFactory
 import kotlin.coroutines.CoroutineContext
 
 // TODO(refactor): Unify creation just like it's done for UserDataBase
@@ -30,18 +30,21 @@ actual class GlobalDatabaseProvider(
     private val database: GlobalDatabase
 
     init {
+        val schema = GlobalDatabase.Schema
         driver = if (encrypt) {
+            System.loadLibrary("sqlcipher")
             AndroidSqliteDriver(
-                schema = GlobalDatabase.Schema,
+                schema = schema,
                 context = context,
                 name = dbName,
-                factory = SupportFactory(passphrase.value)
+                factory = SupportOpenHelperFactory(passphrase.value, null, true)
             )
         } else {
             AndroidSqliteDriver(
-                schema = GlobalDatabase.Schema,
+                schema = schema,
                 context = context,
-                name = dbName
+                name = dbName,
+                callback = SqliteCallback(schema)
             )
         }