--- a/persistence/src/androidMain/kotlin/com/wire/kalium/persistence/db/UserDatabase.kt
+++ b/persistence/src/androidMain/kotlin/com/wire/kalium/persistence/db/UserDatabase.kt
@@ -8,7 +8,7 @@ import com.wire.kalium.persistence.UserDatabase
 import com.wire.kalium.persistence.dao.UserIDEntity
 import com.wire.kalium.persistence.util.FileNameUtil
 import kotlinx.coroutines.CoroutineDispatcher
-import net.sqlcipher.database.SupportFactory
+import net.zetetic.database.sqlcipher.SupportOpenHelperFactory
 
 sealed interface DatabaseCredentials {
     data class Passphrase(val value: UserDBSecret) : DatabaseCredentials
@@ -35,17 +35,19 @@ fun userDatabaseBuilder(
     val dbName = FileNameUtil.userDBName(userId)
 
     val driver: AndroidSqliteDriver = if (encrypt) {
+        System.loadLibrary("sqlcipher")
         AndroidSqliteDriver(
             schema = UserDatabase.Schema,
             context = context,
             name = dbName,
-            factory = SupportFactory(passphrase.value)
+            factory = SupportOpenHelperFactory(passphrase.value, null, true)
         )
     } else {
         AndroidSqliteDriver(
             schema = UserDatabase.Schema,
             context = context,
-            name = dbName
+            name = dbName,
+            callback = SqliteCallback(UserDatabase.Schema)
         )
     }
     val credentials = if (encrypt) {
@@ -66,11 +68,15 @@ fun inMemoryDatabase(
         schema = UserDatabase.Schema,
         context = context,
         name = null,
-        factory = SupportFactory(passphrase)
+        factory = SupportOpenHelperFactory(passphrase)
     )
     return UserDatabaseBuilder(
-        userId, driver, dispatcher, PlatformDatabaseData(
-            context, DatabaseCredentials.Passphrase(
+        userId = userId,
+        sqlDriver = driver,
+        dispatcher = dispatcher,
+        platformDatabaseData = PlatformDatabaseData(
+            context = context,
+            databaseCredentials = DatabaseCredentials.Passphrase(
                 UserDBSecret(passphrase)
             )
         )