--- a/persistence/src/androidMain/kotlin/com/wire/kalium/persistence/db/UserDatabase.kt
+++ b/persistence/src/androidMain/kotlin/com/wire/kalium/persistence/db/UserDatabase.kt
@@ -64,6 +64,7 @@ fun inMemoryDatabase(
     dispatcher: CoroutineDispatcher
 ): UserDatabaseBuilder {
     val passphrase = "testPass".toByteArray()
+    System.loadLibrary("sqlcipher")
     val driver = AndroidSqliteDriver(
         schema = UserDatabase.Schema,
         context = context,