--- a/persistence/src/commonTest/kotlin/com/wire/kalium/persistence/dao/message/MessageDAOBenchmarkTest.kt
+++ b/persistence/src/commonTest/kotlin/com/wire/kalium/persistence/dao/message/MessageDAOBenchmarkTest.kt
@@ -3,13 +3,16 @@ package com.wire.kalium.persistence.dao.message
 import com.wire.kalium.persistence.BaseDatabaseTest
 import com.wire.kalium.persistence.dao.ConversationDAO
 import com.wire.kalium.persistence.dao.UserDAO
-import com.wire.kalium.persistence.dao.UserIDEntity
 import com.wire.kalium.persistence.utils.stubs.newConversationEntity
 import com.wire.kalium.persistence.utils.stubs.newUserEntity
+import com.wire.kalium.util.KaliumDispatcherImpl
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.runTest
 import kotlinx.datetime.Instant
 import kotlin.random.Random
+import kotlin.random.nextInt
 import kotlin.test.BeforeTest
 import kotlin.test.Ignore
 import kotlin.test.Test
@@ -37,9 +40,9 @@ class MessageDAOBenchmarkTest : BaseDatabaseTest() {
         userDAO = db.userDAO
     }
 
-    @OptIn(ExperimentalTime::class)
+    @OptIn(ExperimentalTime::class, ExperimentalCoroutinesApi::class)
     @Test
-    fun insertTextMessages() = runTest {
+    fun insertRandomMessages() = runTest {
         setupData()
         val count = MESSAGE_COUNT
         val messagesToInsert = generateRandomMessages(count)
@@ -47,15 +50,14 @@ class MessageDAOBenchmarkTest : BaseDatabaseTest() {
             messageDAO.insertOrIgnoreMessages(messagesToInsert)
         }
 
-        println("Took $duration to insert $count text messages")
+        println("Took $duration to insert $count random messages")
     }
 
-    @Suppress("LongMethod")
     private fun generateRandomMessages(count: Int): List<MessageEntity> {
         val conversations = listOf(conversationEntity1)
         val users = listOf(userEntity1, userEntity2)
         return buildList {
-            repeat(count / 4) {
+            repeat(count) {
                 add(
                     MessageEntity.Regular(
                         id = it.toString(),
@@ -64,71 +66,7 @@ class MessageDAOBenchmarkTest : BaseDatabaseTest() {
                         senderUserId = users.random().id,
                         status = MessageEntity.Status.values().random(),
                         visibility = MessageEntity.Visibility.values().random(),
-                        content = MessageEntityContent.Text("Text content for message $it"),
-                        senderClientId = Random.nextLong(2_000).toString(),
-                        editStatus = MessageEntity.EditStatus.NotEdited,
-                        senderName = "senderName"
-                    )
-                )
-
-                add(
-                    MessageEntity.System(
-                        id = it.toString(),
-                        conversationId = conversations.random().id,
-                        date = Instant.fromEpochSeconds(it.toLong()),
-                        senderUserId = users.random().id,
-                        status = MessageEntity.Status.values().random(),
-                        visibility = MessageEntity.Visibility.values().random(),
-                        content = MessageEntityContent.MemberChange(
-                            listOf(UserIDEntity("value", "domain")),
-                            MessageEntity.MemberChangeType.REMOVED
-                        ),
-                        senderName = "senderName"
-                    )
-                )
-
-                add(
-                    MessageEntity.Regular(
-                        id = it.toString(),
-                        conversationId = conversations.random().id,
-                        date = Instant.fromEpochSeconds(it.toLong()),
-                        senderUserId = users.random().id,
-                        status = MessageEntity.Status.values().random(),
-                        visibility = MessageEntity.Visibility.values().random(),
-                        content = MessageEntityContent.Asset(
-                            1000,
-                            assetName = "test name",
-                            assetMimeType = "MP4",
-                            assetDownloadStatus = null,
-                            assetOtrKey = byteArrayOf(1),
-                            assetSha256Key = byteArrayOf(1),
-                            assetId = "assetId",
-                            assetToken = "",
-                            assetDomain = "domain",
-                            assetEncryptionAlgorithm = "",
-                            assetWidth = 111,
-                            assetHeight = 111,
-                            assetDurationMs = 10,
-                            assetNormalizedLoudness = byteArrayOf(1),
-                        ),
-                        senderClientId = Random.nextLong(2_000).toString(),
-                        editStatus = MessageEntity.EditStatus.NotEdited,
-                        senderName = "senderName"
-                    )
-                )
-
-                add(
-                    MessageEntity.Regular(
-                        id = it.toString(),
-                        conversationId = conversations.random().id,
-                        date = Instant.fromEpochSeconds(it.toLong()),
-                        senderUserId = users.random().id,
-                        status = MessageEntity.Status.values().random(),
-                        visibility = MessageEntity.Visibility.values().random(),
-                        content = MessageEntityContent.Unknown(
-                            typeName = null,
-                            Random.nextBytes(100000)
-                        ),
+                        content = generateRandomMessageContent(),
                         senderClientId = Random.nextLong(2_000).toString(),
                         editStatus = MessageEntity.EditStatus.NotEdited,
                         senderName = "senderName"
@@ -138,24 +76,80 @@ class MessageDAOBenchmarkTest : BaseDatabaseTest() {
         }
     }
 
-    @OptIn(ExperimentalTime::class)
+    private fun generateRandomMessageContent() = when (Random.nextInt(0..3)) {
+        0 -> MessageEntityContent.Unknown(typeName = null, Random.nextBytes(1000))
+        1 -> MessageEntityContent.Text(Random.nextBytes(100).toString())
+        2 -> MessageEntityContent.Asset(
+            1000,
+            assetName = "test name",
+            assetMimeType = "MP4",
+            assetDownloadStatus = null,
+            assetOtrKey = byteArrayOf(1),
+            assetSha256Key = byteArrayOf(1),
+            assetId = "assetId",
+            assetToken = "",
+            assetDomain = "domain",
+            assetEncryptionAlgorithm = "",
+            assetWidth = 111,
+            assetHeight = 111,
+            assetDurationMs = 10,
+            assetNormalizedLoudness = byteArrayOf(1),
+        )
+
+        else -> MessageEntityContent.Knock(Random.nextBoolean())
+    }
+
+    @OptIn(ExperimentalTime::class, ExperimentalCoroutinesApi::class)
     @Test
-    fun queryTextMessagesNormal() = runTest {
+    fun queryIncreasinglyBiggerAmountByConversationAndVisibility() = runTest {
         setupData()
-        val totalMessageCount = MESSAGE_COUNT
-        val messagesToInsert = generateRandomMessages(totalMessageCount)
-        messageDAO.insertOrIgnoreMessages(messagesToInsert)
-        repeat(4) {
+        var totalMessageCount = MESSAGE_COUNT
+        repeat(3) { count ->
+            val messagesToInsert = generateRandomMessages(totalMessageCount)
+            messageDAO.insertOrIgnoreMessages(messagesToInsert)
             measureTime {
                 messageDAO.getMessagesByConversationAndVisibility(
                     conversationEntity1.id,
                     totalMessageCount,
                     0,
-                    MessageEntity.Visibility.values().toList()
+                    listOf(MessageEntity.Visibility.VISIBLE)
                 ).first()
             }.also {
-                println("Took $it to query $totalMessageCount messages")
+                println("Took $it to query visible messages from a single conversation, with $totalMessageCount random messages inserted")
             }
+            totalMessageCount += MESSAGE_COUNT
+        }
+    }
+
+    @OptIn(ExperimentalTime::class, ExperimentalCoroutinesApi::class)
+    @Test
+    fun concurrentInsertAndQuery() = runTest {
+        setupData()
+        val totalMessageCount = MESSAGE_COUNT
+        val pageSize = 50
+        val messagesToInsert = generateRandomMessages(totalMessageCount)
+        measureTime {
+            val insertingJob = launch(KaliumDispatcherImpl.io) {
+                messagesToInsert.forEach { messageEntity ->
+                    messageDAO.insertOrIgnoreMessage(messageEntity)
+                }
+            }
+            launch(KaliumDispatcherImpl.io) {
+                while (insertingJob.isActive) {
+                    messageDAO.getMessagesByConversationAndVisibility(
+                        conversationId = conversationEntity1.id,
+                        limit = pageSize,
+                        offset = 0,
+                        visibility = listOf(MessageEntity.Visibility.VISIBLE)
+                    ).first()
+                }
+            }.join()
+        }.also {
+            println(
+                """
+                Took $it to insert $totalMessageCount random messages while a worker was constantly querying the database.
+                """.trimIndent()
+            )
         }
     }
 
@@ -166,6 +160,6 @@ class MessageDAOBenchmarkTest : BaseDatabaseTest() {
     }
 
     private companion object {
-        const val MESSAGE_COUNT = 1000
+        const val MESSAGE_COUNT = 1_000
     }
 }