----
#### PR Submission Checklist for internal contributors

- The **PR Title**
  - [X] conforms to the style of semantic commits messages¹ supported in Wire's Github Workflow²
  - [X] answers the question: _If merged, this PR will: ..._ ³

- The **PR Description**
  - [X] is free of optional paragraphs and you have filled the relevant parts to the best of your ability
----

# What's new in this PR?

### Issues

Reloaded spends too much time locking/unlocking the Datbase.

### Causes

In an app as complex as Reloaded, it's expected that we perform multiple read and write operations during long periods of time, especially when processing events.

However, the current SQLCipher build we're using is not great when it comes to optimise locks and unlocks, as explained [here](https://www.zetetic.net/blog/2022/05/11/sqlcipher-android-refresh/).

### Solutions

- Use the new SQLCipher "refresh"
- Enable [WAL Journal Mode](https://www.sqlite.org/wal.html) for the database.
- Pass `journal_mode=wal` to JVM's SQLite Driver.

> **Warning**: Small breaking change on Android. Some internal library has `allowBackup:true` and `supportsRtl:true` on their manifest, so we need to add `tools:replace="android:allowBackup,android:supportsRtl"` to Reloaded's `AndroidManifest.xml`

### Benchmark

Running on an **Android Emulator**:

#### Inserting 5k messages, each message in its own transaction

| Old SQLCipher | New SQLCipher & WAL |
|---------------|---------------------|
| 25.091472500s | 10.675776300s       |


#### Inserting 5_000 random messages, each message in its own transaction, while another thread is constantly querying the last 50 messages

| Old SQLCipher | New SQLCipher & WAL |
|---------------|---------------------|
| 30.714321900s | 14.509833300s       |

Same test, but on JVM (both without SQLCipher)

| WAL OFF       | WAL ON       |
|---------------|--------------|
| 15.552227300s | 5.173206700s | 

It's a huge difference.


### Testing

#### Test Coverage

A new benchmark for concurrent DB read/write.

- [X] I have added automated test to this contribution

----
#### PR Post Merge Checklist for internal contributors

 - [X] If any soft of configuration variable was introduced by this PR, it has been added to the relevant documents and the CI jobs have been updated.
----
##### References
1. https://sparkbox.com/foundry/semantic_commit_messages
1. https://github.com/wireapp/.github#usage
1. E.g. `feat(conversation-list): Sort conversations by most emojis in the title #SQPIT-764`.